Extract base class for `AssignedBeforeEach` and `AssignedAfterEach`
